Changed Get(Async)KeyState return type to short.

+++ src/native/windows/org_lwjgl_opengl_WindowsKeyboard.c
@@ -41,11 +41,11 @@
 #include <jni.h>
 #include "org_lwjgl_opengl_WindowsKeyboard.h"
 
-JNIEXPORT jint JNICALL Java_org_lwjgl_opengl_WindowsKeyboard_GetKeyState(JNIEnv *env, jclass unused, jint virt_key) {
+JNIEXPORT jshort JNICALL Java_org_lwjgl_opengl_WindowsKeyboard_GetKeyState(JNIEnv *env, jclass unused, jint virt_key) {
 	return GetKeyState(virt_key);
 }
 
-JNIEXPORT jint JNICALL Java_org_lwjgl_opengl_WindowsKeyboard_GetAsyncKeyState(JNIEnv *env, jclass unused, jint virt_key) {
+JNIEXPORT jshort JNICALL Java_org_lwjgl_opengl_WindowsKeyboard_GetAsyncKeyState(JNIEnv *env, jclass unused, jint virt_key) {
 	return GetAsyncKeyState(virt_key);
 }
